# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Bayyina CRM API
  description: |
    REST API for Bayyina CRM — an education center management system.

    ## Authentication
    All protected endpoints require a Bearer token in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```
    Tokens are obtained via `POST /api/auth/login` or `POST /api/auth/verify-otp`.
    Tokens expire after **30 days**.

    ## Roles
    | Role      | Description                                      |
    |-----------|--------------------------------------------------|
    | `admin`   | Full access to all resources                     |
    | `teacher` | Read access to students; manage groups/payments  |
    | `student` | Read-only access to own data                     |

    ## Error Format
    All errors follow this structure:
    ```json
    { "code": "errorCode", "message": "Human readable message" }
    ```

  version: 1.0.0
  contact:
    name: Mardonbek Khamidov
    email: mardonbek@bayyina.uz
servers:
  - url: http://localhost:4000/api
    description: Local development
  - url: https://api.bayyina.uz/api
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token issued on login or OTP verification. Valid for 30 days.

  # ─────────────────────────────────────────────
  # Reusable schemas
  # ─────────────────────────────────────────────
  schemas:
...
    CourseType:
      type: object
      description: A configurable course category used to track lead interest (e.g. IELTS / English / Advanced).
      properties:
        _id:
          type: string
          example: "664f1a2b3c4d5e6f7a8b9c10"
        name:
          type: string
          description: Display name
          example: "IELTS Tayyorlov"
        type:
          type: string
          description: Course subject area
          example: "English"
        direction:
          type: string
          description: Level or direction
          example: "Advanced"
        createdBy:
          oneOf:
            - type: string
              nullable: true
            - type: object
              properties:
                _id: { type: string, example: "664f1a2b3c4d5e6f7a8b9c0f" }
                firstName: { type: string, example: "Sarvinoz" }
                lastName: { type: string, example: "Yusupova" }
        createdAt:
          type: string
          format: date-time
          example: "2026-04-01T08:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-04-01T08:00:00.000Z"

    # ── RejectionReason ──────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "invalidCredentials"
        message:
          type: string
          example: "Telefon raqam yoki parol noto'g'ri"

    # ── Record ──────────────────────────────────
...
  parameters:
...
    pageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
      description: Page number (1-based)
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
      description: Number of results per page (default 20, max 100)
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            tokenMissing:
              summary: No token provided
              value: { code: "tokenMissing", message: "Token mavjud emas" }
            tokenExpired:
              summary: Token has expired
              value: { code: "tokenExpired", message: "Token muddati tugagan" }
            invalidToken:
              summary: Token is malformed
              value: { code: "invalidToken", message: "Yaroqsiz token kiritildi" }
    Forbidden:
      description: Authenticated but insufficient role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: "forbidden"
            message: "Ruxsat berilmadi"
...
    BadRequest:
      description: Validation error or bad input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Rate limit exceeded (120 req / 3 min)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: "tooManyRequests"
            message: "Juda ko'p so'rovlar. Iltimos, 3 daqiqadan keyin urinib ko'ring"
...
paths:
...
  /course-types:
    get:
      tags: [Course Types]
      summary: List course types
      description: Returns a paginated list of course types. Supports free-text search (`q`) and optional `type`/`direction` filters. **Admin only.**
      operationId: getCourseTypes
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/limitParam"
        - name: q
          in: query
          description: Search by name, type, or direction
          schema:
            type: string
            example: "IELTS"
        - name: type
          in: query
          description: Filter by course subject area (partial match)
          schema:
            type: string
            example: "English"
        - name: direction
          in: query
          description: Filter by direction/level (partial match)
          schema:
            type: string
            example: "Advanced"
      responses:
        "200":
          description: Course types returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  courseTypes:
                    type: array
                    items:
                      $ref: "#/components/schemas/CourseType"
                  total: { type: integer, example: 8 }
                  page: { type: integer, example: 1 }
                  limit: { type: integer, example: 20 }
                  totalPages: { type: integer, example: 1 }
                  hasNextPage: { type: boolean, example: false }
                  hasPrevPage: { type: boolean, example: false }
                  code: { type: string, example: courseTypesFound }
                  message: { type: string, example: "Kurs turlari ro'yxati olindi" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    post:
      tags: [Course Types]
      summary: Create a course type
      description: Creates a new course type. **Admin only.**
      operationId: createCourseType
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: "IELTS Tayyorlov"
                type:
                  type: string
                  example: "English"
                direction:
                  type: string
                  example: "Advanced"
      responses:
        "201":
          description: Course type created
          content:
            application/json:
              schema:
                type: object
                properties:
                  courseType:
                    $ref: "#/components/schemas/CourseType"
                  code: { type: string, example: courseTypeCreated }
                  message: { type: string, example: "Kurs turi muvaffaqiyatli yaratildi" }
        "400":
          description: Missing name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example: { code: "missingField", message: "Kurs turi nomi kiritilishi shart" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /course-types/{id}:
    get:
      tags: [Course Types]
      summary: Get a course type by ID
      operationId: getCourseType
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "664f1a2b3c4d5e6f7a8b9c10"
      responses:
        "200":
          description: Course type found
          content:
            application/json:
              schema:
                type: object
                properties:
                  courseType:
                    $ref: "#/components/schemas/CourseType"
                  code: { type: string, example: courseTypesFound }
                  message: { type: string, example: "Kurs turlari ro'yxati olindi" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Course type not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example: { code: "courseTypeNotFound", message: "Kurs turi topilmadi" }
        "429":
          $ref: "#/components/responses/TooManyRequests"

    put:
      tags: [Course Types]
      summary: Update a course type
      operationId: updateCourseType
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "664f1a2b3c4d5e6f7a8b9c10"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "General English"
                type:
                  type: string
                  example: "English"
                direction:
                  type: string
                  example: "Beginner"
      responses:
        "200":
          description: Course type updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  courseType:
                    $ref: "#/components/schemas/CourseType"
                  code: { type: string, example: courseTypeUpdated }
                  message: { type: string, example: "Kurs turi muvaffaqiyatli yangilandi" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Course type not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example: { code: "courseTypeNotFound", message: "Kurs turi topilmadi" }
        "429":
          $ref: "#/components/responses/TooManyRequests"

    delete:
      tags: [Course Types]
      summary: Delete a course type
      operationId: deleteCourseType
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "664f1a2b3c4d5e6f7a8b9c10"
      responses:
        "200":
          description: Course type deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: string, example: courseTypeDeleted }
                  message: { type: string, example: "Kurs turi muvaffaqiyatli o'chirildi" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Course type not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example: { code: "courseTypeNotFound", message: "Kurs turi topilmadi" }
        "429":
          $ref: "#/components/responses/TooManyRequests"

  # ══════════════════════════════════════════
  # REJECTION REASONS
  # ══════════════════════════════════════════
...